+++ tests/mongodb.js
@@ -1560,7 +1560,6 @@ describe('MongoDB', () => {
 		});
 
 		it('Should throw if connection to DB fails', async () => {
-
 			const collection = stubMongo(false);
 
 			const mongodb = new MongoDB(config);
@@ -1762,6 +1761,569 @@ describe('MongoDB', () => {
 
 			sinon.assert.notCalled(collection);
 			sinon.assert.notCalled(findAndModify);
+		});
+	});
+
+	describe('getIndexes()', () => {
+
+		it('Should throw if no model is passed', async () => {
+			const mongodb = new MongoDB(config);
+			await assert.rejects(() => mongodb.getIndexes(null), {
+				code: MongoDBError.codes.INVALID_MODEL
+			});
+		});
+
+		it('Should throw if connection to DB fails', async () => {
+
+			const collectionStub = stubMongo(false);
+
+			const mongodb = new MongoDB(config);
+
+			await assert.rejects(() => mongodb.getIndexes(getModel()), {
+				message: 'Error getting DB',
+				code: MongoDBError.codes.MONGODB_INTERNAL_ERROR
+			});
+
+			sinon.assert.notCalled(collectionStub);
+		});
+
+		it('Should throw if mongodb indexes method fails', async () => {
+
+			const indexesStub = sinon.stub().rejects(new Error('Indexes internal error'));
+			const collectionStub = stubMongo(true, { indexes: indexesStub });
+
+			const mongodb = new MongoDB(config);
+
+			await assert.rejects(() => mongodb.getIndexes(getModel()), {
+				message: 'Indexes internal error',
+				code: MongoDBError.codes.MONGODB_INTERNAL_ERROR
+			});
+
+			sinon.assert.calledOnce(collectionStub);
+			sinon.assert.calledWithExactly(collectionStub, 'myCollection');
+		});
+
+		it('Should return the indexes from the model collection', async () => {
+
+			const indexesStub = sinon.stub().resolves([
+				{
+					v: 2,
+					key: { field: 1 },
+					name: 'some-index',
+					ns: 'myDatabase.myCollection'
+				},
+				{
+					v: 2,
+					key: { uniqueField: 1 },
+					name: 'unique-index',
+					ns: 'myDatabase.myCollection',
+					unique: true
+				}
+			]);
+
+			const collectionStub = stubMongo(true, { indexes: indexesStub });
+
+			const mongodb = new MongoDB(config);
+			const model = getModel();
+
+			const result = await mongodb.getIndexes(model);
+
+			assert.deepStrictEqual(result, [
+				{
+					name: 'some-index',
+					key: { field: 1 },
+					unique: false
+				},
+				{
+					name: 'unique-index',
+					key: { uniqueField: 1 },
+					unique: true
+				}
+			]);
+
+			sinon.assert.calledOnce(collectionStub);
+			sinon.assert.calledWithExactly(collectionStub, 'myCollection');
+
+			sinon.assert.calledOnce(indexesStub);
+			sinon.assert.calledWithExactly(indexesStub);
+		});
+	});
+
+	describe('createIndexes()', () => {
+
+		const indexes = [
+			{
+				name: 'some-index',
+				key: {
+					field: 1
+				}
+			},
+			{
+				name: 'unique-index',
+				key: {
+					uniqueField: 1
+				},
+				unique: true
+			}
+		];
+
+		it('Should throw if no model is passed', async () => {
+			const mongodb = new MongoDB(config);
+			await assert.rejects(() => mongodb.createIndexes(null), {
+				code: MongoDBError.codes.INVALID_MODEL
+			});
+		});
+
+		it('Should throw if connection to DB fails', async () => {
+			const collectionStub = stubMongo(false);
+
+			const mongodb = new MongoDB(config);
+
+			await assert.rejects(() => mongodb.createIndexes(getModel(), indexes), {
+				message: 'Error getting DB',
+				code: MongoDBError.codes.MONGODB_INTERNAL_ERROR
+			});
+
+			sinon.assert.notCalled(collectionStub);
+		});
+
+		it('Should throw if mongodb createIndexes method fails', async () => {
+
+			const createIndexesStub = sinon.stub().rejects(new Error('createIndexes internal error'));
+			const collectionStub = stubMongo(true, { createIndexes: createIndexesStub });
+
+			const mongodb = new MongoDB(config);
+
+			await assert.rejects(() => mongodb.createIndexes(getModel(), indexes), {
+				message: 'createIndexes internal error',
+				code: MongoDBError.codes.MONGODB_INTERNAL_ERROR
+			});
+
+			sinon.assert.calledOnce(collectionStub);
+			sinon.assert.calledWithExactly(collectionStub, 'myCollection');
+		});
+
+		[
+
+			null,
+			undefined,
+			'not an object',
+			['array'],
+			{},
+			{
+				key: 'not an object'
+			},
+			{
+				name: { not: 'a string' },
+				key: { field: 1 }
+			},
+			{
+				name: 'some-index',
+				key: { field: 1 },
+				unique: 'not a boolean'
+			}
+
+		].forEach(index => {
+
+			it('Should throw if received indexes are invalid', async () => {
+
+				const mongodb = new MongoDB(config);
+				const model = getModel();
+
+				await assert.rejects(() => mongodb.createIndexes(model, [index]), {
+					name: 'MongoDBError',
+					code: MongoDBError.codes.INVALID_INDEX
+				});
+			});
+		});
+
+		it('Should throw if received indexes are invalid', async () => {
+
+			const mongodb = new MongoDB(config);
+			const model = getModel();
+
+			await assert.rejects(() => mongodb.createIndexes(model, { invalid: 'index' }));
+
+		});
+
+		it('Should return true if can create the indexes into the model collection successfully', async () => {
+
+			const createIndexesStub = sinon.stub().resolves({ ok: true });
+			const collectionStub = stubMongo(true, { createIndexes: createIndexesStub });
+
+			const mongodb = new MongoDB(config);
+			const model = getModel();
+
+			const result = await mongodb.createIndexes(model, indexes);
+
+			assert.deepStrictEqual(result, true);
+
+			sinon.assert.calledOnce(collectionStub);
+			sinon.assert.calledWithExactly(collectionStub, 'myCollection');
+
+			sinon.assert.calledOnce(createIndexesStub);
+			sinon.assert.calledWithExactly(createIndexesStub, indexes);
+		});
+
+		it('Should return false if can\'t create the indexes into the model collection', async () => {
+
+			const createIndexesStub = sinon.stub().resolves({ ok: false });
+			const collectionStub = stubMongo(true, { createIndexes: createIndexesStub });
+
+			const mongodb = new MongoDB(config);
+			const model = getModel();
+
+			const result = await mongodb.createIndexes(model, indexes);
+
+			assert.deepStrictEqual(result, false);
+
+			sinon.assert.calledOnce(collectionStub);
+			sinon.assert.calledWithExactly(collectionStub, 'myCollection');
+
+			sinon.assert.calledOnce(createIndexesStub);
+			sinon.assert.calledWithExactly(createIndexesStub, indexes);
+		});
+	});
+
+	describe('createIndex()', () => {
+
+		const index = {
+			name: 'some-index',
+			key: {
+				field: 1
+			}
+		};
+
+		it('Should throw if no model is passed', async () => {
+			const mongodb = new MongoDB(config);
+			await assert.rejects(() => mongodb.createIndex(null), {
+				code: MongoDBError.codes.INVALID_MODEL
+			});
+		});
+
+		it('Should throw if connection to DB fails', async () => {
+
+			const collectionStub = stubMongo(false);
+
+			const mongodb = new MongoDB(config);
+
+			await assert.rejects(() => mongodb.createIndex(getModel(), index), {
+				message: 'Error getting DB',
+				code: MongoDBError.codes.MONGODB_INTERNAL_ERROR
+			});
+
+			sinon.assert.notCalled(collectionStub);
+		});
+
+		it('Should throw if mongodb createIndexes method fails', async () => {
+
+			const createIndexesStub = sinon.stub().rejects(new Error('createIndexes internal error'));
+			const collectionStub = stubMongo(true, { createIndexes: createIndexesStub });
+
+			const mongodb = new MongoDB(config);
+
+			await assert.rejects(() => mongodb.createIndex(getModel(), index), {
+				message: 'createIndexes internal error',
+				code: MongoDBError.codes.MONGODB_INTERNAL_ERROR
+			});
+
+			sinon.assert.calledOnce(collectionStub);
+			sinon.assert.calledWithExactly(collectionStub, 'myCollection');
+		});
+
+		[
+
+			null,
+			undefined,
+			'not an object',
+			['array'],
+			{},
+			{
+				key: 'not an object'
+			},
+			{
+				name: { not: 'a string' },
+				key: { field: 1 }
+			},
+			{
+				name: 'some-index',
+				key: { field: 1 },
+				unique: 'not a boolean'
+			}
+
+		].forEach(invalidIndex => {
+
+			it('Should throw if received index is invalid', async () => {
+
+				sinon.spy(MongoDB.prototype, 'createIndexes');
+
+				const mongodb = new MongoDB(config);
+				const model = getModel();
+
+				await assert.rejects(() => mongodb.createIndex(model, invalidIndex), {
+					name: 'MongoDBError',
+					code: MongoDBError.codes.INVALID_INDEX
+				});
+
+				sinon.assert.calledOnce(MongoDB.prototype.createIndexes);
+				sinon.assert.calledWithExactly(MongoDB.prototype.createIndexes, model, [invalidIndex]);
+			});
+		});
+
+		it('Should return true if can create the index into the model collection successfully', async () => {
+
+			sinon.stub(MongoDB.prototype, 'createIndexes')
+				.resolves(true);
+
+			const mongodb = new MongoDB(config);
+			const model = getModel();
+
+			const result = await mongodb.createIndex(model, index);
+
+			assert.deepStrictEqual(result, true);
+
+			sinon.assert.calledOnce(MongoDB.prototype.createIndexes);
+			sinon.assert.calledWithExactly(MongoDB.prototype.createIndexes, model, [index]);
+		});
+
+		it('Should return false if can\'t create the indexes into the model collection', async () => {
+
+			sinon.stub(MongoDB.prototype, 'createIndexes')
+				.resolves(false);
+
+			const mongodb = new MongoDB(config);
+			const model = getModel();
+
+			const result = await mongodb.createIndex(model, index);
+
+			assert.deepStrictEqual(result, false);
+
+			sinon.assert.calledOnce(MongoDB.prototype.createIndexes);
+			sinon.assert.calledWithExactly(MongoDB.prototype.createIndexes, model, [index]);
+		});
+	});
+
+	describe('dropIndex()', () => {
+
+		const index = 'some-index';
+
+		it('Should throw if no model is passed', async () => {
+			const mongodb = new MongoDB(config);
+			await assert.rejects(() => mongodb.dropIndex(null), {
+				code: MongoDBError.codes.INVALID_MODEL
+			});
+		});
+
+		it('Should throw if connection to DB fails', async () => {
+
+			const collectionStub = stubMongo(false);
+
+			const mongodb = new MongoDB(config);
+
+			await assert.rejects(() => mongodb.dropIndex(getModel(), index), {
+				message: 'Error getting DB',
+				code: MongoDBError.codes.MONGODB_INTERNAL_ERROR
+			});
+
+			sinon.assert.notCalled(collectionStub);
+		});
+
+		it('Should throw if mongodb dropIndex method fails', async () => {
+
+			const dropIndexStub = sinon.stub().rejects(new Error('dropIndex internal error'));
+			const collectionStub = stubMongo(true, { dropIndex: dropIndexStub });
+
+			const mongodb = new MongoDB(config);
+
+			await assert.rejects(() => mongodb.dropIndex(getModel(), index), {
+				message: 'dropIndex internal error',
+				code: MongoDBError.codes.MONGODB_INTERNAL_ERROR
+			});
+
+			sinon.assert.calledOnce(collectionStub);
+			sinon.assert.calledWithExactly(collectionStub, 'myCollection');
+		});
+
+		[
+
+			null,
+			undefined,
+			1,
+			['array'],
+			{},
+			{
+				key: 'not an object'
+			},
+			{
+				name: { not: 'a string' },
+				key: { field: 1 }
+			},
+			{
+				name: 'some-index',
+				key: { field: 1 },
+				unique: 'not a boolean'
+			}
+
+		].forEach(invalidIndex => {
+
+			it('Should throw if received index is invalid', async () => {
+
+				sinon.spy(MongoDB.prototype, 'createIndexes');
+
+				const mongodb = new MongoDB(config);
+				const model = getModel();
+
+				await assert.rejects(() => mongodb.dropIndex(model, invalidIndex), {
+					name: 'MongoDBError',
+					code: MongoDBError.codes.INVALID_INDEX
+				});
+			});
+		});
+
+		it('Should return true if can drop the index from the model collection successfully', async () => {
+
+			const dropIndexStub = sinon.stub().resolves({ ok: true });
+			const collectionStub = stubMongo(true, { dropIndex: dropIndexStub });
+
+			const mongodb = new MongoDB(config);
+			const model = getModel();
+
+			const result = await mongodb.dropIndex(model, index);
+
+			assert.deepStrictEqual(result, true);
+
+			sinon.assert.calledOnce(collectionStub);
+			sinon.assert.calledWithExactly(collectionStub, 'myCollection');
+
+			sinon.assert.calledOnce(dropIndexStub);
+			sinon.assert.calledWithExactly(dropIndexStub, index);
+		});
+
+		it('Should return false if can\'t drop the index from the model collection', async () => {
+
+			const dropIndexStub = sinon.stub().resolves({ ok: false });
+			const collectionStub = stubMongo(true, { dropIndex: dropIndexStub });
+
+			const mongodb = new MongoDB(config);
+			const model = getModel();
+
+			const result = await mongodb.dropIndex(model, index);
+
+			assert.deepStrictEqual(result, false);
+
+			sinon.assert.calledOnce(collectionStub);
+			sinon.assert.calledWithExactly(collectionStub, 'myCollection');
+
+			sinon.assert.calledOnce(dropIndexStub);
+			sinon.assert.calledWithExactly(dropIndexStub, index);
+		});
+	});
+
+	describe('dropIndexes()', () => {
+
+		const indexes = ['some-index', 'other-index'];
+
+		it('Should throw if no model is passed', async () => {
+			const mongodb = new MongoDB(config);
+			await assert.rejects(() => mongodb.dropIndexes(null), {
+				code: MongoDBError.codes.INVALID_MODEL
+			});
+		});
+
+		it('Should call dropIndex for each received indexName and return true when all the received indexes was dropped', async () => {
+
+			const dropIndexStub = sinon.stub().resolves({ ok: true });
+			const collectionStub = stubMongo(true, { dropIndex: dropIndexStub });
+
+			const mongodb = new MongoDB(config);
+			const model = getModel();
+
+			const result = await mongodb.dropIndexes(model, indexes);
+
+			assert.deepStrictEqual(result, true);
+
+			sinon.assert.calledTwice(collectionStub);
+			sinon.assert.calledWithExactly(collectionStub.getCall(1), 'myCollection');
+			sinon.assert.calledWithExactly(collectionStub.getCall(0), 'myCollection');
+
+			sinon.assert.calledTwice(dropIndexStub);
+
+			indexes.forEach((index, i) => {
+				sinon.assert.calledWithExactly(dropIndexStub.getCall(i), index);
+			});
+		});
+
+		it('Should return false when some of the indexes drop operation returns false', async () => {
+
+			const dropIndexStub = sinon.stub();
+
+			dropIndexStub.onFirstCall()
+				.resolves({ ok: false });
+
+			dropIndexStub.onSecondCall()
+				.resolves({ ok: true });
+
+			const collectionStub = stubMongo(true, { dropIndex: dropIndexStub });
+
+			const mongodb = new MongoDB(config);
+			const model = getModel();
+
+			const result = await mongodb.dropIndexes(model, indexes);
+
+			assert.deepStrictEqual(result, false);
+
+			sinon.assert.calledTwice(collectionStub);
+			sinon.assert.calledWithExactly(collectionStub.getCall(1), 'myCollection');
+			sinon.assert.calledWithExactly(collectionStub.getCall(0), 'myCollection');
+
+			sinon.assert.calledTwice(dropIndexStub);
+
+			indexes.forEach((index, i) => {
+				sinon.assert.calledWithExactly(dropIndexStub.getCall(i), index);
+			});
+		});
+
+		it('Should throw when the dropIndex method rejects', async () => {
+
+			const dropIndexStub = sinon.stub().rejects();
+			const collectionStub = stubMongo(true, { dropIndex: dropIndexStub });
+
+			const mongodb = new MongoDB(config);
+			const model = getModel();
+
+			await assert.rejects(() => mongodb.dropIndexes(model, indexes));
+
+			sinon.assert.calledTwice(collectionStub);
+			sinon.assert.calledWithExactly(collectionStub.getCall(1), 'myCollection');
+			sinon.assert.calledWithExactly(collectionStub.getCall(0), 'myCollection');
+
+			sinon.assert.calledTwice(dropIndexStub);
+
+			indexes.forEach((index, i) => {
+				sinon.assert.calledWithExactly(dropIndexStub.getCall(i), index);
+			});
+		});
+
+		[
+
+			null,
+			undefined,
+			1,
+			'not an array',
+			{}
+
+		].forEach(invalidIndexes => {
+
+			it('Should throw if received index is invalid', async () => {
+
+				const mongodb = new MongoDB(config);
+				const model = getModel();
+
+				await assert.rejects(() => mongodb.dropIndexes(model, invalidIndexes), {
+					name: 'MongoDBError',
+					code: MongoDBError.codes.INVALID_INDEX
+				});
+			});
 		});
 	});
 });