+++ CHANGELOG.md
@@ -5,9 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.9.0] - 2020-01-13
 ### Added
 - `increment` method
+- `getIndexes` method for getting the collection indexes
+- `createIndex` method for creating an index into the collection
+- `createIndexes` method for creating multiple indexes into the collection
+- `dropIndex` method for dropping an index from the collection
+- `dropIndexes` method for dropping multiple indexes from the collection
 
 ## [1.8.7] - 2019-12-31
 ### Fixed